Reduce overall compile time by ~10% - variable.h
Reduces the number of template instantiations of the Variable
polymorphism. This was being instantiated 108 times previously, which
impacts both compile and link time.
Locally, a clang-10 debug build sped up from ~11 minutes to ~10 minutes.
